Fixing banner anchor underline for mobile browsers
Fixing banner anchor underline for mobile browsers while in landscape.
Up until now, the underline method use is that of Medium.com, which is
a background gradient hack, with the goal of a more precise visual
underline.

It doesn’t work well in certain instances on mobile browsers, and it’s
a lot of code, to which I don’t think is worth having for what it does.

I’ve instated the simple, traditional underline, since it works
everywhere.